+++ src/components/Form/NicknameInput.jsx
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { chekcDuplicateNickname } from '@apis/axios/auth';
-import InputErrorBox from '../Form/InputErrorBox';
-import CustomInput from './CustomInput';
+import InputErrorBox from './InputErrorBox';
+import CustomInput from '../Input/CustomInput';
 
 const NicknameInput = ({ register, errors }) => {
   return (
