+++ src/components/Form/PasswordConfirmInput.jsx
@@ -1,5 +1,5 @@
-import InputErrorBox from '../Form/InputErrorBox';
-import CustomInput from './CustomInput';
+import InputErrorBox from './InputErrorBox';
+import CustomInput from '../Input/CustomInput';
 
 const PasswordConfirmInput = ({ register, errors, getValues }) => {
   return (
